+++ school-medical-management/src/main/java/com/swp391/school_medical_management/modules/users/controllers/NurseController.java
@@ -1,9 +1,11 @@
 package com.swp391.school_medical_management.modules.users.controllers;
 
-import java.time.LocalDate;
-import java.util.List;
-import java.util.Map;
-
+import com.swp391.school_medical_management.modules.users.dtos.request.*;
+import com.swp391.school_medical_management.modules.users.dtos.response.*;
+import com.swp391.school_medical_management.modules.users.entities.MedicalRequestEntity;
+import com.swp391.school_medical_management.modules.users.entities.UserEntity.UserRole;
+import com.swp391.school_medical_management.modules.users.services.impl.NurseService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
@@ -11,41 +13,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
-import com.swp391.school_medical_management.modules.users.dtos.request.BlogRequest;
-import com.swp391.school_medical_management.modules.users.dtos.request.HealthCheckResultRequest;
-import com.swp391.school_medical_management.modules.users.dtos.request.MedicalEventRequest;
-import com.swp391.school_medical_management.modules.users.dtos.request.ReplyFeedbackRequest;
-import com.swp391.school_medical_management.modules.users.dtos.request.UpdateMedicalRequestStatus;
-import com.swp391.school_medical_management.modules.users.dtos.request.VaccineResultRequest;
-import com.swp391.school_medical_management.modules.users.dtos.response.BlogResponse;
-import com.swp391.school_medical_management.modules.users.dtos.response.ClassStudentDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.FeedbackDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.HealthCheckFormDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.HealthCheckResultDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.MedicalEventDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.MedicalRecordDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.MedicalRequestDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.MedicalRequestDetailDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.OnGoingProgramDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.StudentDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.UserDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.VaccineFormDTO;
-import com.swp391.school_medical_management.modules.users.dtos.response.VaccineResultDTO;
-import com.swp391.school_medical_management.modules.users.entities.MedicalRequestEntity;
-import com.swp391.school_medical_management.modules.users.entities.UserEntity.UserRole;
-import com.swp391.school_medical_management.modules.users.services.impl.NurseService;
-
-import jakarta.validation.Valid;
+import java.time.LocalDate;
+import java.util.List;
 
 @Validated
 @RestController
@@ -82,20 +53,20 @@ public class NurseController {
 
     @PutMapping("/{requestId}/status")
     public ResponseEntity<MedicalRequestDTO> updateMedicalRequestStatus(@PathVariable int requestId,
-            @RequestBody @Valid UpdateMedicalRequestStatus request) {
+                                                                        @RequestBody @Valid UpdateMedicalRequestStatus request) {
         MedicalRequestDTO medicalRequestDTO = nurseService.updateMedicalRequestStatus(requestId, request);
         return ResponseEntity.ok(medicalRequestDTO);
     }
 
     @GetMapping("/health-check-forms/{healthCheckFormId}")
-    public ResponseEntity<HealthCheckFormDTO> getHealthCheckFormById(@PathVariable Long healthCheckFormId) {
+    public ResponseEntity<HealthCheckFormDTO> getHealthCheckFormById(@PathVariable int healthCheckFormId) {
         HealthCheckFormDTO healthCheckFormDTO = nurseService.getHealthCheckFormById(healthCheckFormId);
         return ResponseEntity.ok(healthCheckFormDTO);
     }
 
     @GetMapping("/health-check-forms/program/{programId}")
     public ResponseEntity<List<HealthCheckFormDTO>> getHealthCheckFormsByProgram(
-            @PathVariable Long programId,
+            @PathVariable int programId,
             @RequestParam(required = false) Boolean committed) {
         List<HealthCheckFormDTO> list = nurseService.getHealthCheckFormsByProgram(programId, committed);
         return ResponseEntity.ok(list);
@@ -103,26 +74,26 @@ public class NurseController {
 
     @GetMapping("/vaccine-forms/program/{programId}")
     public ResponseEntity<List<VaccineFormDTO>> getVaccineFormsByProgram(
-            @PathVariable Long programId,
+            @PathVariable int programId,
             @RequestParam(required = false) Boolean committed) {
         List<VaccineFormDTO> forms = nurseService.getVaccineFormsByProgram(programId, committed);
         return ResponseEntity.ok(forms);
     }
 
     @GetMapping("/vaccine-forms/{vaccineFormId}")
-    public ResponseEntity<VaccineFormDTO> getVaccineFormById(@PathVariable Long vaccineFormId) {
+    public ResponseEntity<VaccineFormDTO> getVaccineFormById(@PathVariable int vaccineFormId) {
         VaccineFormDTO vaccineFormDTO = nurseService.getVaccinFormById(vaccineFormId);
         return ResponseEntity.ok(vaccineFormDTO);
     }
 
     @GetMapping("/vaccine-forms/not-send/program/{programId}")
-    public ResponseEntity<List<VaccineFormDTO>> getNotSentVaccineFormsByProgram(@PathVariable Long programId) {
+    public ResponseEntity<List<VaccineFormDTO>> getNotSentVaccineFormsByProgram(@PathVariable int programId) {
         List<VaccineFormDTO> forms = nurseService.getNotSentVaccineFormsByProgram(programId);
         return ResponseEntity.ok(forms);
     }
 
     @GetMapping("/health-check-forms/not-send/program/{programId}")
-    public ResponseEntity<List<HealthCheckFormDTO>> getNotSentHealthCheckFormsByProgram(@PathVariable Long programId) {
+    public ResponseEntity<List<HealthCheckFormDTO>> getNotSentHealthCheckFormsByProgram(@PathVariable int programId) {
         List<HealthCheckFormDTO> forms = nurseService.getNotSentHealthCheckFormsByProgram(programId);
         return ResponseEntity.ok(forms);
     }
@@ -133,13 +104,13 @@ public class NurseController {
     }
 
     // @GetMapping("/draft-form/count/{programId}")
-    // public ResponseEntity<Map<String, Long>>
-    // countDraftFormsByProgram(@PathVariable Long programId) {
+    // public ResponseEntity<Map<String, int>>
+    // countDraftFormsByProgram(@PathVariable int programId) {
     // return ResponseEntity.ok(nurseService.countDraftFormByProgram(programId));
     // }
 
     @GetMapping("/medical-events/{studentId}")
-    public ResponseEntity<List<MedicalEventDTO>> getMedicalEventsByStudent(@PathVariable Long studentId) {
+    public ResponseEntity<List<MedicalEventDTO>> getMedicalEventsByStudent(@PathVariable int studentId) {
         List<MedicalEventDTO> events = nurseService.getMedicalEventsByStudent(studentId);
         return ResponseEntity.ok(events);
     }
@@ -147,21 +118,21 @@ public class NurseController {
     @PostMapping("/medical-event")
     public ResponseEntity<MedicalEventDTO> createMedicalEvent(@RequestBody MedicalEventRequest request) {
         String nurseId = SecurityContextHolder.getContext().getAuthentication().getName();
-        MedicalEventDTO medicalEventDTO = nurseService.createMedicalEvent(Long.parseLong(nurseId), request);
+        MedicalEventDTO medicalEventDTO = nurseService.createMedicalEvent(Integer.parseInt(nurseId), request);
         return ResponseEntity.status(HttpStatus.CREATED).body(medicalEventDTO);
     }
 
     @PutMapping("/medical-event/{medicalEventId}")
-    public ResponseEntity<MedicalEventDTO> updateMedicalEvent(@PathVariable Long medicalEventId,
-            @RequestBody MedicalEventRequest request) {
+    public ResponseEntity<MedicalEventDTO> updateMedicalEvent(@PathVariable int medicalEventId,
+                                                              @RequestBody MedicalEventRequest request) {
         String nurseId = SecurityContextHolder.getContext().getAuthentication().getName();
-        MedicalEventDTO medicalEventDTO = nurseService.updateMedicalEvent(Long.parseLong(nurseId), medicalEventId,
+        MedicalEventDTO medicalEventDTO = nurseService.updateMedicalEvent(Integer.parseInt(nurseId), medicalEventId,
                 request);
         return ResponseEntity.ok(medicalEventDTO);
     }
 
     @GetMapping("/medical-event/{medicalEventId}")
-    public ResponseEntity<MedicalEventDTO> getMedicalEvent(@PathVariable Long medicalEventId) {
+    public ResponseEntity<MedicalEventDTO> getMedicalEvent(@PathVariable int medicalEventId) {
         MedicalEventDTO medicalEventDTO = nurseService.getMedicalEvent(medicalEventId);
         return ResponseEntity.ok(medicalEventDTO);
     }
@@ -173,7 +144,7 @@ public class NurseController {
     }
 
     @DeleteMapping("/medical-event/{medicalEventId}")
-    public ResponseEntity<Void> deleteMedicalEvent(@PathVariable Long medicalEventId) {
+    public ResponseEntity<Void> deleteMedicalEvent(@PathVariable int medicalEventId) {
         nurseService.deleteMedicalEvent(medicalEventId);
         return ResponseEntity.noContent().build();
     }
@@ -186,26 +157,26 @@ public class NurseController {
     }
 
     @PutMapping("/health-check-result/{healCheckResultId}")
-    public ResponseEntity<HealthCheckResultDTO> putMethodName(@PathVariable Long healCheckResultId,
-            @RequestBody HealthCheckResultRequest request) {
+    public ResponseEntity<HealthCheckResultDTO> putMethodName(@PathVariable int healCheckResultId,
+                                                              @RequestBody HealthCheckResultRequest request) {
         HealthCheckResultDTO healthCheckResultDTO = nurseService.updateHealthCheckResult(healCheckResultId, request);
         return ResponseEntity.ok(healthCheckResultDTO);
     }
 
     @GetMapping("/health-check-result/{healCheckResultId}")
-    public ResponseEntity<HealthCheckResultDTO> getHealthCheckResult(@PathVariable Long healCheckResultId) {
+    public ResponseEntity<HealthCheckResultDTO> getHealthCheckResult(@PathVariable int healCheckResultId) {
         HealthCheckResultDTO healthCheckResultDTO = nurseService.getHealthCheckResult(healCheckResultId);
         return ResponseEntity.ok(healthCheckResultDTO);
     }
 
     @GetMapping("/health-check-result/program/{programId}")
-    public ResponseEntity<List<HealthCheckResultDTO>> getHealthCheckResultByProgram(@PathVariable Long programId) {
+    public ResponseEntity<List<HealthCheckResultDTO>> getHealthCheckResultByProgram(@PathVariable int programId) {
         List<HealthCheckResultDTO> resultDTOList = nurseService.getHealthCheckResultByProgram(programId);
         return ResponseEntity.ok(resultDTOList);
     }
 
     @DeleteMapping("/health-check-result/{healCheckResultId}")
-    public ResponseEntity<Void> deleteHealthCheckResult(@PathVariable Long healCheckResultId) {
+    public ResponseEntity<Void> deleteHealthCheckResult(@PathVariable int healCheckResultId) {
         nurseService.deleteHealthCheckResult(healCheckResultId);
         return ResponseEntity.noContent().build();
     }
@@ -217,26 +188,26 @@ public class NurseController {
     }
 
     @PutMapping("/vaccine-result/{vaccineResultId}")
-    public ResponseEntity<VaccineResultDTO> updateVaccineResult(@PathVariable Long vaccineResultId,
-            @RequestBody VaccineResultRequest request) {
+    public ResponseEntity<VaccineResultDTO> updateVaccineResult(@PathVariable int vaccineResultId,
+                                                                @RequestBody VaccineResultRequest request) {
         VaccineResultDTO vaccineResultDTO = nurseService.updateVaccineResult(vaccineResultId, request);
         return ResponseEntity.ok(vaccineResultDTO);
     }
 
     @GetMapping("/vaccine-result/{vaccineResultId}")
-    public ResponseEntity<VaccineResultDTO> getVaccineResult(@PathVariable Long vaccineResultId) {
+    public ResponseEntity<VaccineResultDTO> getVaccineResult(@PathVariable int vaccineResultId) {
         VaccineResultDTO vaccineResultDTO = nurseService.getVaccineResult(vaccineResultId);
         return ResponseEntity.ok(vaccineResultDTO);
     }
 
     @GetMapping("/vaccine-result/program/{programId}")
-    public ResponseEntity<List<VaccineResultDTO>> getVaccineResultByProgram(@PathVariable Long programId) {
+    public ResponseEntity<List<VaccineResultDTO>> getVaccineResultByProgram(@PathVariable int programId) {
         List<VaccineResultDTO> results = nurseService.getVaccineResultByProgram(programId);
         return ResponseEntity.ok(results);
     }
 
     @DeleteMapping("/vaccine-result/{vaccineResultId}")
-    public ResponseEntity<Void> deleteVaccineResult(@PathVariable Long vaccineResultId) {
+    public ResponseEntity<Void> deleteVaccineResult(@PathVariable int vaccineResultId) {
         nurseService.deleteVaccineResult(vaccineResultId);
         return ResponseEntity.noContent().build();
     }
@@ -274,8 +245,8 @@ public class NurseController {
     // từng tham gia chương trình vaccine
     @GetMapping("/students/not-vaccinated")
     public ResponseEntity<List<StudentDTO>> getStudentsNotVaccinated(
-            @RequestParam(required = false) Long vaccineProgramId,
-            @RequestParam(required = false) Long vaccineName) {
+            @RequestParam(required = false) int vaccineProgramId,
+            @RequestParam(required = false) int vaccineName) {
         List<StudentDTO> result = nurseService.getStudentsNotVaccinated(vaccineProgramId, vaccineName);
         return ResponseEntity.ok(result);
     }
@@ -299,14 +270,14 @@ public class NurseController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<BlogResponse> updatePost(@PathVariable Long id,
-            @RequestBody BlogRequest request) {
+    public ResponseEntity<BlogResponse> updatePost(@PathVariable int id,
+                                                   @RequestBody BlogRequest request) {
         BlogResponse response = nurseService.update(id, request);
         return ResponseEntity.ok(response);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletePost(@PathVariable Long id) {
+    public ResponseEntity<Void> deletePost(@PathVariable int id) {
         nurseService.delete(id);
         return ResponseEntity.noContent().build();
     }
@@ -319,7 +290,7 @@ public class NurseController {
     }
 
     @GetMapping("/class/{classId}")
-    public ResponseEntity<List<StudentDTO>> getStudentsByClass(@PathVariable Long classId) {
+    public ResponseEntity<List<StudentDTO>> getStudentsByClass(@PathVariable int classId) {
         List<StudentDTO> students = nurseService.getStudentsByClass(classId);
         return ResponseEntity.ok(students);
     }
@@ -343,7 +314,7 @@ public class NurseController {
     }
 
     @GetMapping("/class/{classId}/students")
-    public ResponseEntity<List<MedicalRequestDTO>> getByClass(@PathVariable Long classId) {
+    public ResponseEntity<List<MedicalRequestDTO>> getByClass(@PathVariable int classId) {
         return ResponseEntity.ok(nurseService.getByClass(classId));
     }
 
@@ -355,33 +326,33 @@ public class NurseController {
     }
 
     @PostMapping("/create-health-check-form/{programId}")
-    public ResponseEntity<String> createHealthCheckForm(@PathVariable Long programId) {
+    public ResponseEntity<String> createHealthCheckForm(@PathVariable int programId) {
         nurseService.createFormsForHealthCheckProgram(programId);
         return ResponseEntity.ok("Forms generated successfully for all students.");
     }
 
     @PostMapping("/create-vaccine-form/{programId}")
-    public ResponseEntity<String> createVaccineForm(@PathVariable Long programId) {
+    public ResponseEntity<String> createVaccineForm(@PathVariable int programId) {
         nurseService.createFormsForVaccineProgram(programId);
         return ResponseEntity.ok("Vaccine forms generated successfully.");
     }
 
     // @PostMapping("/create-default-by-program/{programId}")
     // public ResponseEntity<List<HealthCheckResultDTO>>
-    // createResultsByProgram(@PathVariable Long programId) {
+    // createResultsByProgram(@PathVariable int programId) {
     // List<HealthCheckResultDTO> results = nurseService
     // .createDefaultHealthCheckResultsByProgramId(programId);
     // return ResponseEntity.ok(results);
     // }
 
     @PostMapping("/create-healthCheckResult-byProgram-/{programId}")
-    public ResponseEntity<List<HealthCheckResultDTO>> createResultsByProgramId(@PathVariable Long programId) {
+    public ResponseEntity<List<HealthCheckResultDTO>> createResultsByProgramId(@PathVariable int programId) {
         List<HealthCheckResultDTO> createdResults = nurseService.createResultsByProgramId(programId);
         return ResponseEntity.ok(createdResults);
     }
 
     @PostMapping("/create-vaccineResults-byProgram/{programId}")
-    public ResponseEntity<List<VaccineResultDTO>> createVaccineResultsByProgramId(@PathVariable Long programId) {
+    public ResponseEntity<List<VaccineResultDTO>> createVaccineResultsByProgramId(@PathVariable int programId) {
         List<VaccineResultDTO> createdResults = nurseService.createVaccineResultsByProgramId(programId);
         return ResponseEntity.ok(createdResults);
     }
